Remove openopt, allow spaces in $BROWSER
Remove the openopt variable, instead opting to allow spaces in `$BROWSER`.

`$BROWSER` as used by the XDG implementation allows spaces (and `%s`
printf-style formatting strings), and is actually a PATH-style list.
While this patch doesn't make git-open compliant with the XDG-style
variable, it does allow for the use of `BROWSER="powershell.exe
Start"` in the environment.